requirements.txt
discord.py==1.7.3
pyTelegramBotAPI==4.4.0
python-dotenv==0.20.0
